+++ sample/sample6-global-middlewares/AllErrorsHandler.ts
@@ -1,8 +1,8 @@
 import {ServerResponse, ServerRequest} from "http";
-import {Middleware, GlobalMiddleware} from "../../src/decorator/decorators";
+import {Middleware, MiddlewareGlobalBefore} from "../../src/decorator/decorators";
 import {ErrorHandlerInterface} from "../../src/middleware/ErrorHandlerInterface";
 
-@GlobalMiddleware()
+@MiddlewareGlobalBefore()
 export class AllErrorsHandler implements ErrorHandlerInterface {
 
     error(error: any, request: any, response: any, next?: Function): void {