+++ sample/sample6-global-middlewares/CompressionMiddleware.ts
@@ -1,8 +1,8 @@
 import {ServerResponse, ServerRequest} from "http";
-import {GlobalMiddleware} from "../../src/decorator/decorators";
+import {MiddlewareGlobalBefore} from "../../src/decorator/decorators";
 import {MiddlewareInterface} from "../../src/middleware/MiddlewareInterface";
 
-@GlobalMiddleware()
+@MiddlewareGlobalBefore()
 export class CompressionMiddleware implements MiddlewareInterface {
 
     use(request: any, response: any, next?: Function): any {