+++ sample/sample6-global-middlewares/EndTimerMiddleware.ts
@@ -1,8 +1,7 @@
-import {ServerResponse, ServerRequest} from "http";
-import {Middleware, GlobalMiddleware} from "../../src/decorator/decorators";
+import {MiddlewareGlobalAfter} from "../../src/decorator/decorators";
 import {MiddlewareInterface} from "../../src/middleware/MiddlewareInterface";
 
-@GlobalMiddleware({ afterAction: true })
+@MiddlewareGlobalAfter()
 export class EndTimerMiddleware implements MiddlewareInterface {
 
     use(request: any, response: any, next?: Function): any {