+++ sample/sample6-global-middlewares/LoggerMiddleware.ts
@@ -1,8 +1,8 @@
 import {ServerResponse, ServerRequest} from "http";
-import {Middleware, GlobalMiddleware} from "../../src/decorator/decorators";
+import {Middleware, MiddlewareGlobalBefore} from "../../src/decorator/decorators";
 import {MiddlewareInterface} from "../../src/middleware/MiddlewareInterface";
 
-@GlobalMiddleware()
+@MiddlewareGlobalBefore()
 export class LoggerMiddleware implements MiddlewareInterface {
 
     use(request: any, response: any, next?: Function): any {